+++ src/client/widgets/common/widget.js
@@ -1065,13 +1065,7 @@ class Widget extends EventEmitter {
 
     isVisible() {
 
-        return this.getProp('visible') && this.parent.childrenVisible()
-
-    }
-
-    childrenVisible() {
-
-        return this.isVisible()
+        return this.getProp('visible') && this.parent.isVisible()
 
     }