+++ src/launcher/toolbar.js
@@ -3,6 +3,7 @@ var {remote, ipcRenderer} = require('electron'),
     menu = new Menu(),
     terminal = require('./terminal'),
     settings = require('./settings'),
+    keyboardJS = require('keyboardjs/dist/keyboard.min.js'),
     {midilist, openDocs} = remote.getGlobal('launcherSharedGlobals'),
     serverStarted = false,
     serverStart = ()=>{
@@ -16,18 +17,21 @@ var {remote, ipcRenderer} = require('electron'),
 
 var start = new MenuItem({
     label: 'Start',
-    click: serverStart
+    click: serverStart,
+    accelerator: 'f5'
 })
 var stop = new MenuItem({
     label: 'Stop',
-    click: serverStop
+    click: serverStop,
+    accelerator: 'f6'
 })
 var newWindow = new MenuItem({
     label: 'New window',
     visible: false,
     click: ()=>{
         ipcRenderer.send('openClient')
-    }
+    },
+    accelerator: 'CmdOrCtrl + n'
 })
 menu.append(start)
 menu.append(stop)
@@ -39,19 +43,22 @@ menu.append(new MenuItem({
     label: 'Load',
     click: ()=>{
         settings.load()
-    }
+    },
+    accelerator: 'CmdOrCtrl + o'
 }))
 menu.append(new MenuItem({
     label: 'Save',
     click: ()=>{
         settings.save()
-    }
+    },
+    accelerator: 'CmdOrCtrl + s'
 }))
 menu.append(new MenuItem({
     label: 'Save as...',
     click: ()=>{
         settings.saveAs()
-    }
+    },
+    accelerator: 'CmdOrCtrl + shift + s'
 }))
 menu.append(new MenuItem({
     type: 'separator'
@@ -60,14 +67,16 @@ menu.append(new MenuItem({
     label: 'List MIDI Devices',
     click: ()=>{
         midilist()
-    }
+    },
+    accelerator: 'CmdOrCtrl + m'
 }))
 menu.append(new MenuItem({type: 'submenu' , label: 'Console', submenu: [
     new MenuItem({
         label: 'Clear',
         click: ()=>{
             terminal.clear()
-        }
+        },
+        accelerator: 'CmdOrCtrl + l'
     }),
     new MenuItem({
         label: 'Autoscroll',
@@ -105,6 +114,29 @@ menu.append(new MenuItem({
 menu.append(new MenuItem({
     role: 'Quit'
 }))
+
+
+function bindShortcuts(menu) {
+
+    for (let m of menu.items) {
+
+        if (m.type === 'submenu') {
+
+            bindShortcuts(m.submenu)
+
+        } else if (m.accelerator && m.click) {
+
+            keyboardJS.bind(m.accelerator.replace('CmdOrCtrl', 'mod'), (e)=>{
+                m.click()
+            })
+
+        }
+
+    }
+
+}
+
+bindShortcuts(menu)
 
 class Toolbar {