+++ CHANGELOG.md
@@ -1,14 +1,41 @@
 # Changelog
 
-# 1.8.9
+## next
+
+- bug fixes
+  - switch: widget not reacting at first touch when traversing is enabled on parent
+
+- custom module
+  - expose `process` global
+
+- scripting
+  - add `openUrl` function
+
+## 1.8.10
+
+- bug fixes
+  - modal: issues with children's visibility
+  - console: allow multiple arguments in console.log()
+
+- ui
+  - launcher: add many keyboard shortcuts
+  - client: add keyboard shortcuts for clearing the console
+
+- scripting
+  - add `setFocus` function to focus an input widget programmatically
+
+- misc
+  - add `usePercents` client option
+
+## 1.8.9
 
 - bug fixes
   - script: issue when using the `options` argument in `set()` (options leaked to subsequent set() calls in the script)
    - multixy: `decimals` property not applied
    - multixy: spring behavior not working until all points are released  
-   - custom module: prevent require() from reading submodules files each time and instead return the object in memory 
+   - custom module: prevent require() from reading submodules files each time and instead return the object in memory
 
-# 1.8.8
+## 1.8.8
 
 - launcher
   - add file browser button for the `theme` option and fix parsing path containing spaces if only one theme is set